MILESTONE: project rearranged and TOPIC SEARCH added
The whole project structure has been re-arranged to provide a more
cleaned, radable, reusable, ooriented version of the whole software.
It works as the previous version, plu the important adding of TOPIC
SEARCH

Topic is a new way to classify tweets based on their hashtags and
keywords.
Untill now, it is roughtly raw and simple, but it coul be dramatically
improved in future updates!